api: Move unique key helpers to generate-keys.ts
